requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.11
django-allauth==65.14.3
django-filter==25.2
idna==3.11
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
